+++ react/CheckoutB2BContext.tsx
@@ -1,25 +1,32 @@
 import React from 'react'
+import { withToast } from 'vtex.styleguide'
+
+import type { WithToast } from './typings'
 
 type CheckoutB2BContextData = {
   pending: boolean
   setPending: React.Dispatch<React.SetStateAction<boolean>>
+  showToast: WithToast['showToast']
 }
 
 const CheckoutB2BContext = React.createContext<CheckoutB2BContextData | null>(
   null
 )
 
-export function CheckoutB2BProvider({
+function CheckoutB2BProviderWrapper({
   children,
-}: React.PropsWithChildren<unknown>) {
+  showToast,
+}: React.PropsWithChildren<WithToast>) {
   const [pending, setPending] = React.useState(false)
 
   return (
-    <CheckoutB2BContext.Provider value={{ pending, setPending }}>
+    <CheckoutB2BContext.Provider value={{ pending, setPending, showToast }}>
       {children}
     </CheckoutB2BContext.Provider>
   )
 }
+
+export const CheckoutB2BProvider = withToast(CheckoutB2BProviderWrapper)
 
 export function useCheckoutB2BContext() {
   const context = React.useContext(CheckoutB2BContext)