+++ react/hooks/useToast.ts
@@ -1,0 +1,7 @@
+import { useCheckoutB2BContext } from '../CheckoutB2BContext'
+
+export function useToast() {
+  const { showToast } = useCheckoutB2BContext()
+
+  return showToast
+}